requirements.txt
numpy==1.13.3
tqdm==4.18.0
pillow==4.2.1
sklearn==0.20.2
imagehash==4.0
keras==2.2.4